Refactor: Ensure consistent typing of cup-related variables
Tracks are now officially limited to 32768 and therefore the cups per
list are limited to 8192 (these limits are not yet enforced in the tool)

+++ src/payload/game/ui/page/RaceMenuPage.cpp
@@ -7,7 +7,7 @@
 
 // RaceMenuPage::onButtonSelect() patch
 // Get the next track (for VS mode)
-kmCallDefCpp(0x8085AE80, u32, int raceNum) {
+kmCallDefCpp(0x8085AE80, u32, u16 raceNum) {
     const u16 track = SectionManager::instance->globalContext->trackOrder[raceNum];
     CupManager::currentSzs = track;
     return CupData::tracks[track].specialSlot;
@@ -15,7 +15,7 @@ kmCallDefCpp(0x8085AE80, u32, int raceNum) {
 
 // RaceMenuPage::onButtonSelect() patch
 // Get the next arena (for BT mode)
-kmCallDefCpp(0x8085AF30, u32, int raceNum) {
+kmCallDefCpp(0x8085AF30, u32, u16 raceNum) {
 
     // Account for tiebreakers
     raceNum = raceNum % CupManager::GetTrackCount(true);
